+++ tests/testthat/test-fc_request.R
@@ -1,6 +1,7 @@
 context("fc_request")
 
 testthat::skip_on_cran()
+testthat::skip_if(!fc_ping())
 
 test_that("fc_request fails with invalid credentials", {
   expect_error(fc_request("GET", "/user",
@@ -12,9 +13,9 @@ test_that("fc_request fails with missing credentials", {
   skip_if(Sys.getenv("FC_PRIVATE_KEY") == "")
 
   expect_error(fc_request("GET", "/user", public_key = ""),
-               regexp = "public_key is empty")
+               regexp = "public_key missing")
   expect_error(fc_request("GET", "/user", private_key = ""),
-               regexp = "private_key is empty")
+               regexp = "private_key missing")
 })
 
 test_that("fc_request work", {