+++ tests/testthat/test-fc_wrappers.R
@@ -1,7 +1,8 @@
 context("wrappers")
 
 testthat::skip_on_cran()
-skip_if(Sys.getenv("FC_PRIVATE_KEY") == "")
+testthat::skip_if(Sys.getenv("FC_PRIVATE_KEY") == "")
+testthat::skip_if(!fc_ping())
 
 
 test_that("fc_get_user() works", {
@@ -40,12 +41,3 @@ test_that("fc_get_data_range() works", {
   expect_true(names(data)[[1]] == "sensors")
 })
 
-test_that("fc_get_data_range() handles 204s", {
-
-  data <- fc_get_data_range(
-    station_id = stations[[1]]$station_name,
-    data_group = "raw",
-    from = as.integer(as.POSIXct(stations[[1]]$db$min_date) - 100),
-    to =  as.integer(as.POSIXct(stations[[1]]$db$min_date) - 10))
-  expect_equal(data, NULL)
-})