+++ NEWS.md
@@ -1,6 +1,8 @@
 # rfieldclimate 0.1.1
 
 * Maintenance release
+* fc_request gained timeout argument
+* new function fc_ping to test api ping
 * Fail gracefully if API has errors or is not available
 
 # rfieldclimate 0.1.0